LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/state-after-navigate-restore-expected.txt



FAIL entry.getState() is properly restored after cross-document back assert_equals: expected 3 but got 1
